// src/models/enums/PlayerStates.ts
export enum PlayerStates {
    DISPLAY_RESULTS = 'DISPLAY_RESULTS',
    IN_GAME = 'IN_GAME',
    IN_GOLDEN_BALLS_ROUND = 'IN_GOLDEN_BALLS_ROUND',
    NOT_IN_GAME = 'NOT_IN_GAME',
    PREPARING_GOLDEN_BALLS = 'PREPARING_GOLDEN_BALLS',
    SEARCHING_FOR_GAME = 'SEARCHING_FOR_GAME',
    WAITING_FOR_GOLDEN_BALLS = 'WAITING_FOR_GOLDEN_BALLS',
}